move proxy and UUID stuff to their own files
