+++ openpathsampling/experimental/simstore/serialization_helpers.py
@@ -9,6 +9,11 @@ from .tools import flatten_all, nested_update, group_by_function
 from .tools import is_iterable, is_mappable, is_numpy_iterable
 from . import tools
 from .class_lookup import is_storage_iterable, is_storage_mappable
+from .proxy import GenericLazyLoader
+from .uuids import (
+    has_uuid, get_uuid, set_uuid, encode_uuid, decode_uuid, encoded_uuid_re,
+    is_uuid_string
+)
 
 # UUID recognition and encoding #####################################
 # Things in here might be modified for performance optimization. In
@@ -18,43 +23,6 @@ import sys
 if sys.version_info > (3, ):
     unicode = str
     long = int
-
-def has_uuid(obj):
-    return hasattr(obj, '__uuid__')
-
-
-def get_uuid(obj):
-    # TODO: I can come up with a better string encoding than this
-    try:
-        return str(obj.__uuid__)
-    except AttributeError as e:
-        if obj is None:
-            return obj
-        else:
-            raise e
-
-def set_uuid(obj, uuid):
-    obj.__uuid__ = long(uuid)
-
-
-def encode_uuid(uuid):
-    return "UUID(" + str(uuid) + ")"
-
-
-def decode_uuid(uuid_str):
-    return uuid_str[5:-1]
-
-
-# use the regular expression when looking through an entire JSON string; use
-# the is_uuid_string method for individual objects
-encoded_uuid_re = re.compile("UUID\((?P<uuid>[\-]?[0-9]+)\)")
-
-
-def is_uuid_string(obj):
-    return (
-        isinstance(obj, (str, unicode))
-        and obj[:5] == 'UUID(' and obj[-1] == ')'
-    )
 
 
 # Getting the list of UUIDs bsed on initial objets ###################
@@ -165,6 +133,10 @@ def get_all_uuids(initial_object, known_uuids=None, class_info=None):
         # found_objs += collections.Counter(o.__class__.__name__)
                                           # for o in objects)
         for obj in objects:
+            # TODO: this might be slow; check performance
+            if isinstance(obj, GenericLazyLoader):
+                obj = obj.load()
+
             # TODO: find a way to ensure that objects doesn't go over
             # duplicates here; see lprofile of default_find_uuids to see how
             # often abort due to being in cache in there, and whether we
